refactor(states): factories.go からデバッグメニューと Choice ヘルパを分離する
factories.go が state ファクトリに加え、約190行のデバッグメニュー本体と Choice.Run
ヘルパまで抱え、名と実体がずれていた。デバッグメニューの選択肢とスポーン補助を
debug_menu.go へ、pushChoice・popAfter・stayAfter・pushMessage を Choice の定義先で
ある choice.go へ移す。挙動は変えず、factories.go は state ファクトリに絞る。
